chore: move dir entry and stat to client.py
